Removed DLNA functionality.
Removed DLNA functionality, this was due to the excessive delay trying to navigate a share (at least in mediatomb).
Moved to a browser model instead.
Implemented samba mounting in OS.
Added ability to debug specific classes.
Added sorting of directory lists.
Added delays to tickers so they pause at the start and end.